% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comboTable.R
\name{comboTable}
\alias{comboTable}
\title{Execute Complex Table Queries}
\usage{
comboTable(query_name)
}
\arguments{
\item{query_name}{The name of the query to execute. Type 'ComboQueries$' and press TAB to see available queries}
}
\value{
A dataframe containing the query results
}
\description{
Execute Complex Table Queries
}
\examples{
\dontrun{
# Type ComboQueries$ and press TAB to see available queries
data <- comboTable(ComboQueries$last_filled)
}
}
