% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_domino_connect.R
\name{hr_domino_connect}
\alias{hr_domino_connect}
\title{Connect to HR Domino using environment variables
Username and password will be requested if they are not found}
\usage{
hr_domino_connect()
}
\value{
A connection to the AS400 called 'hr_domino'
}
\description{
Connect to HR Domino using environment variables
Username and password will be requested if they are not found
}
