% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwdhere.R
\name{setwdhere}
\alias{setwdhere}
\title{Running this function (eventually) will set your working directory to the location of the currently open file
Do not use this function in scripts that will be pasted into PowerBI
Right now, it just reminds you of the function}
\usage{
setwdhere()
}
\description{
Running this function (eventually) will set your working directory to the location of the currently open file
Do not use this function in scripts that will be pasted into PowerBI
Right now, it just reminds you of the function
}
