% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\docType{data}
\name{valid_tables}
\alias{valid_tables}
\title{Valid tables and their corresponding schemas}
\format{
An object of class \code{character} of length 10.
}
\usage{
valid_tables
}
\description{
Valid tables and their corresponding schemas
}
\keyword{datasets}
